extract interface from pallet
