+++ cli/src/chain_spec.rs
@@ -28,7 +28,7 @@ use node_runtime::{
     ImOnlineConfig, IndicesConfig, SessionConfig, SessionKeys, SocietyConfig, StakerStatus,
     StakingConfig, SudoConfig, SystemConfig, TechnicalCommitteeConfig,
 };
-use pallet_credit::{CreditData, CreditLevel, CreditSetting};
+use node_primitives::credit::{CreditSetting,CreditLevel,CreditData};
 use pallet_im_online::sr25519::AuthorityId as ImOnlineId;
 use sc_chain_spec::ChainSpecExtension;
 use sc_service::ChainType;
