+++ pallets/deeper-node/src/lib.rs
@@ -57,31 +57,6 @@ impl<AccountId: Decode, BlockNumber: Default> Default for Node<AccountId, BlockN
     }
 }
 
-pub trait NodeInterface<AccountId, BlockNumber> {
-    /// This function tells if the device has been offline for a day
-    fn get_onboard_time(account_id: &AccountId) -> Option<BlockNumber>;
-
-    /// This function tells if the device has ever been online
-    fn im_ever_online(account_id: &AccountId) -> bool;
-
-    /// This function returns how many eras the device has been offline
-    fn get_eras_offline(account_id: &AccountId) -> u32;
-}
-
-impl<AccountId, BlockNumber> NodeInterface<AccountId, BlockNumber> for () {
-    fn get_onboard_time(_account_id: &AccountId) -> Option<BlockNumber> {
-        None
-    }
-
-    fn im_ever_online(_account_id: &AccountId) -> bool {
-        true
-    }
-
-    fn get_eras_offline(_account_id: &AccountId) -> u32 {
-        0
-    }
-}
-
 #[frame_support::pallet]
 pub mod pallet {
     use super::*;
@@ -90,7 +65,7 @@ pub mod pallet {
     use frame_support::{dispatch::DispatchResultWithPostInfo, pallet_prelude::*};
     use frame_system::pallet_prelude::*;
     use frame_system::{self, ensure_signed};
-    use node_primitives::VerifySignatureInterface;
+    use node_primitives::{deeper_node::NodeInterface,VerifySignatureInterface};
     use sp_std::convert::TryInto;
 
     /// Configure the pallet by specifying the parameters and types on which it depends.
