+++ pallets/precompile-credit/src/lib.rs
@@ -29,8 +29,8 @@ use fp_evm::{
 };
 use frame_support::dispatch::{Dispatchable, GetDispatchInfo, PostDispatchInfo};
 use pallet_credit::Call as CreditCall;
-use pallet_credit::CreditInterface;
 use pallet_evm::AddressMapping;
+use node_primitives::credit::CreditInterface;
 
 use crate::util::{Gasometer, RuntimeHelper};
 use alloc::{borrow::ToOwned, vec};
