+++ primitives/Cargo.toml
@@ -17,6 +17,9 @@ frame-system = { default-features = false, git = "https://github.com/paritytech/
 sp-application-crypto = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.19" }
 sp-core = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.19" }
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.19" }
+frame-support = { default-features = false, git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.19"}
+serde = { version = "1.0.101", optional = true, features = ["derive"] }
+enumflags2 = { version = "0.7.4" }
 
 [dev-dependencies]
 sp-serializer = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.19"}
@@ -31,4 +34,5 @@ std = [
 	"sp-application-crypto/std",
 	"sp-core/std",
 	"sp-runtime/std",
+	"serde",
 ]
